fix(Home): primary-background height shrinks in debug mode
Issue:

When the browser is switched into debug mode, the debug panel at the bottom will let the height of the viewport decrease. The height of .primary-background will become exactly the value of the browser height minus the debug panel height, which means that the scroll bar will not appear. But this will let the images be but off.

Solution:
1. Height of .primary-background should not be 100vh. This line should be deleted. Because this line will cause the scrollbar to disappear.
2. Min-height of .primary-background should be 100vh.
3. MIn-height of .street should be specified.